feat(sharding): 集成百度UID生成器并优化数据源配置-未完成1
- 引入百度UID生成器依赖，实现全局唯一ID生成
- 更新ShardingSphere版本至5.4.1，适配新版API
- 重构数据源配置类，简化动态数据源管理
- 移除冗余的自定义数据源路由逻辑
- 调整实体类主键类型为Long，统一ID类型处理
- 更新MyBatis Plus版本至3.5.3.2
- 添加UID生成器Holder模式支持
- 优化分片配置文件结构，提升可维护性
- 清理废弃的分片算法类和无用配置项
- 修复README文档描述不准确问题

+++ src/main/java/com/lolkt/demo/controller/CustomerController.java
@@ -37,7 +37,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}")
-    public CustomerPO selectById(@PathVariable Integer id) {
+    public CustomerPO selectById(@PathVariable  Long id) {
         return customerService.selectById(id);
     }
